+++ src/api/objects/toot.ts
@@ -60,7 +60,6 @@ import {
     type KeysOfValueType,
     type MastodonTag,
     type ScoreType,
-    type StatusList,
     type TagWithUsageCounts,
     type TootLike,
     type TootNumberProp,
@@ -1111,18 +1110,18 @@ export const tootedAt = (toot: TootLike): Date => new Date(toot.createdAt);
 /**
  * Get the earliest toot from a list.
  * @private
- * @param {StatusList} toots - List of toots.
+ * @param {TootLike[]} toots - List of toots.
  * @returns {TootLike | null}
  */
-export const earliestToot = (toots: StatusList): TootLike | null => sortByCreatedAt(toots)[0];
+export const earliestToot = (toots: TootLike[]): TootLike | null => sortByCreatedAt(toots)[0];
 
 /**
  * Get the most recent toot from a list.
  * @private
- * @param {StatusList} toots - List of toots.
+ * @param {TootLike[]} toots - List of toots.
  * @returns {TootLike | null}
  */
-export const mostRecentToot = (toots: StatusList): TootLike | null => sortByCreatedAt(toots).slice(-1)[0];
+export const mostRecentToot = (toots: TootLike[]): TootLike | null => sortByCreatedAt(toots).slice(-1)[0];
 
 /**
  * Returns array with oldest toot first.
@@ -1131,17 +1130,17 @@ export const mostRecentToot = (toots: StatusList): TootLike | null => sortByCrea
  * @param {T} toots - List of toots.
  * @returns {T}
  */
-export function sortByCreatedAt<T extends StatusList>(toots: T): T {
+export function sortByCreatedAt<T extends TootLike[]>(toots: T): T {
     return toots.toSorted((a, b) => (a.createdAt < b.createdAt) ? -1 : 1) as T;
 };
 
 /**
  * Get the Date of the earliest toot in a list.
  * @private
- * @param {StatusList} toots - List of toots.
+ * @param {TootLike[]} toots - List of toots.
  * @returns {Date | null}
  */
-export const earliestTootedAt = (toots: StatusList): Date | null => {
+export const earliestTootedAt = (toots: TootLike[]): Date | null => {
     const earliest = earliestToot(toots);
     return earliest ? tootedAt(earliest) : null;
 };
@@ -1149,10 +1148,10 @@ export const earliestTootedAt = (toots: StatusList): Date | null => {
 /**
  * Get the Date of the most recent toot in a list.
  * @private
- * @param {StatusList} toots - List of toots.
+ * @param {TootLike[]} toots - List of toots.
  * @returns {Date | null}
  */
-export const mostRecentTootedAt = (toots: StatusList): Date | null => {
+export const mostRecentTootedAt = (toots: TootLike[]): Date | null => {
     const newest = mostRecentToot(toots);
     return newest ? tootedAt(newest) : null;
 };