+++ src/types.ts
@@ -33,7 +33,6 @@ export type FeedFetcher = (api: mastodon.rest.Client) => Promise<Toot[]>;
 export type FilterProperty = BooleanFilterName | TootNumberProp;
 export type OptionalNumber = number | null | undefined;
 export type OptionalString = string | null | undefined;
-export type StatusList = TootLike[];
 export type StringSet = Set<string | undefined>;
 export type TootLike = mastodon.v1.Status | SerializableToot | Toot;
 export type TootNumberProp = KeysOfValueType<Toot, number>;
@@ -98,24 +97,19 @@ export type KeysOfValueType<T, SuperClass> = Exclude<
 //    Mastodon API    //
 ////////////////////////
 
-export type CacheableApiObj = (
-    MastodonApiObj[] |
-    MastodonInstances
-);
-
 /**
  * Union type representing any object that can be returned from the Mastodon API and handled by the app
  * in addition to our local extensions like Toot, Account, and TagWithUsageCounts.
  */
-export type MastodonApiObj = (
-    MastodonObjWithID |
+export type ApiObj = (
+    ApiObjWithID |
     MastodonTag |
     mastodon.v1.TrendLink |
     string
 );
 
-// All these types have an id property
-export type MastodonObjWithID = (
+/** Most (but not all) Mastodon API objects have an 'id' property. */
+export type ApiObjWithID = (
     Account |
     TootLike |
     mastodon.v1.Account |
@@ -124,18 +118,28 @@ export type MastodonObjWithID = (
     mastodon.v2.Filter
 );
 
-/**
- * Local extension to the Mastodon Instance type that adds some additional properties
- * @extends {mastodon.v2.Instance}
- * @property {number} [followedPctOfMAU] - Pct of the instance's monthly active users (MAU) the user follows
- * @property {number} [MAU] - Monthly active users of the instance, if available
- */
+/** Any CacheableApiObj will also be written to localForage with these properties. */
+export interface CacheTimestamp {
+    isStale: boolean;
+    updatedAt: Date;
+};
+
+/** ApiObjs are stored in cache as arrays; MastodonInstances is our custom data structure. */
+export type CacheableApiObj = (
+    ApiObj[]
+  | MastodonInstances
+);
+
+/** Local extension to the Mastodon Instance type that adds some additional properties */
 export interface MastodonInstance extends mastodon.v2.Instance {
     followedPctOfMAU?: number;
     MAU?: number;  // MAU data is buried in the Instance hierarchy so this just a copy on the top level
 };
 
-export type MastodonTag = mastodon.v1.Tag | TagWithUsageCounts;
+export type MastodonTag = (
+    TagWithUsageCounts
+  | mastodon.v1.Tag
+);
 
 export interface MinMax {
     min: number;
@@ -153,7 +157,7 @@ export type MinMaxID = {
     max: string;
 };
 
-// Abstract interface for objects that have numToots of some kind
+/** Abstract interface for objects that have numToots of some kind */
 export interface NamedTootCount extends TootCount {
     displayName?: string;
     displayNameWithEmoji?: string; // TODO: just testing this
@@ -183,7 +187,7 @@ export type ScoreType = keyof WeightedScore;
 export type StorableObj = (
     CacheableApiObj |
     FeedFilterSettingsSerialized |
-    MastodonApiObj |
+    ApiObj |
     StringNumberDict |
     Weights |
     number